[refactor] auth 이전
