+++ src/main/java/com/example/finalproject/domain/auth/controller/UserController.java
@@ -1,7 +1,7 @@
-package com.example.finalproject.auth.controller;
+package com.example.finalproject.domain.auth.controller;
 
-import com.example.finalproject.auth.dto.SignupRequestDto;
-import com.example.finalproject.auth.service.UserService;
+import com.example.finalproject.domain.auth.dto.SignupRequestDto;
+import com.example.finalproject.domain.auth.service.UserService;
 import com.example.finalproject.global.responsedto.ApiResponse;
 import com.example.finalproject.global.utils.ResponseUtils;
 import jakarta.servlet.http.HttpServletRequest;
