+++ src/main/java/com/example/finalproject/domain/auth/entity/User.java
@@ -1,6 +1,6 @@
-package com.example.finalproject.auth.entity;
+package com.example.finalproject.domain.auth.entity;
 
-import com.example.finalproject.auth.dto.SignupRequestDto;
+import com.example.finalproject.domain.auth.dto.SignupRequestDto;
 import com.example.finalproject.domain.purchases.dto.MypageRequestDto;
 import com.example.finalproject.global.enums.UserRoleEnum;
 import jakarta.persistence.*;
