+++ src/main/java/com/example/finalproject/domain/auth/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.finalproject.auth.repository;
+package com.example.finalproject.domain.auth.repository;
 
-import com.example.finalproject.auth.entity.User;
+import com.example.finalproject.domain.auth.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
